"Actions" class cleaned up and refactored.
